requirements.txt
python 3.10
OpenCV
cvzone
cv2
google.generativeai
PIL
streamlit
HandDetector
numpy
Image